Modified muttrc to use sidebar-patch
Added mutt sidebar patch to my mutt installation. I also started
to use IMAP folders for the same reason, with the eventual goal
being to have it set up so that I can use the address book in mutt
for both talking to users I've emailed before and for chatting with
IBM employees.