Set shell for vim to bash
I realized that fugitive, the git wrapper expects the shell to
be either bash or zsh. It releys on a shell command that doesn't
exist in fish, my interactive shell of choice so I have to make
sure that vim doesn't do the default and call commands in fish or
the wonders of the fugitive plugin don't work and you get this
lovely blank screen.