Altered the vimrc to better handle swp and backup
I'm getting irritated with swp and backup files poluting the directories
that I'm working in. However, I've been learning over time they exist
for a reason and that turning them off can occationally be a very
foolish thing to do. So, now in an attempt to be more adult I'm moving
to storing them in my .vim directory.