Changes to mutt to better handle html emails
Due to IBM switching to Verse for email and the default setting for
verse being html instead of plaintext I'm now getting a lot more mail
covered in html. Given that the settings for changing verse to plaintext
are hidden far in the depths of verse config, it's just easier to handle
things on my end.

NOTE: My change to handling html now requires that lynx be installed on
the system I'm running mutt on. I need to consider adding this into my
setup scripts.